No longer clears practice cache when there are no storage permissions

+++ app/src/main/java/com/memory_athlete/memoryassistant/preferences/ClearPracticePreference.java
@@ -1,16 +1,20 @@
 package com.memory_athlete.memoryassistant.preferences;
 
 import android.content.Context;
+import android.content.pm.PackageManager;
 import android.content.res.Resources;
-import android.util.AttributeSet;
-
 import android.preference.Preference;
+import android.util.AttributeSet;
 
 import com.memory_athlete.memoryassistant.Helper;
 import com.memory_athlete.memoryassistant.R;
 
 import java.io.File;
 import java.util.Objects;
+
+import static android.Manifest.permission.READ_EXTERNAL_STORAGE;
+import static android.Manifest.permission.WRITE_EXTERNAL_STORAGE;
+import static androidx.core.content.ContextCompat.checkSelfPermission;
 
 /**
  * Created by Manik on 10/09/17.
@@ -49,6 +53,9 @@ public class ClearPracticePreference extends Preference {
     }
 
     private void deleteRecursive(File fileOrDirectory) {
+        if (!(checkSelfPermission(getContext(), READ_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) ||
+                !(checkSelfPermission(getContext(), WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED))
+            return;
         if (fileOrDirectory.isDirectory())
             for (File child : Objects.requireNonNull(fileOrDirectory.listFiles()))
                 deleteRecursive(child);
